Zoom now works
Zooms out when entering the level and zooms in while in the level.

Removed the "Camera" script and replaced it with a script called "CameraScript" as "Camera" should be a protected class and having a script with that name meant that the "Camera" component could not be accessed. Also replaced"MainCamera" with "NewCamera" so you might need to adjust the settings on the camera to suit you.

Added a new gameobject "ZoomOut" which determines the area in which you can zoom out. Also added a new script to it called "ZoomOut".

Added two new coroutines to "CameraScript" that allow us to zoom in and out. These are called by public functions "ZoomIn" and "ZoomOut" which are called by the "ZoomOut" script.